only create aof file and it's manifesto if it doesnt exist

+++ src/aof.rs
@@ -17,13 +17,17 @@ pub fn init_aof_if_enabled(config: &Config) {
 
         // Create the Append-Only File
         let aof_filename = format!("{}.1.incr.aof", appendfilename);
-        let _ = fs::File::create(&path.join(&aof_filename));
+        if !path.join(&aof_filename).exists() {
+            let _ = fs::File::create(&path.join(&aof_filename));
+        }
 
         // Create the manifest File
         let manifest_filename = format!("{}.manifest", appendfilename);
-        let mut f = fs::File::create(&path.join(manifest_filename)).unwrap();
-        // and write
-        let _ = f.write_all(format!("file {} seq 1 type i", &aof_filename).as_bytes());
+        if !path.join(&manifest_filename).exists() {
+            let mut f = fs::File::create(&path.join(manifest_filename)).unwrap();
+            // and write
+            let _ = f.write_all(format!("file {} seq 1 type i", &aof_filename).as_bytes());
+        }
     }
 }
